feat: 로또 당첨번호 입력받는 기능
사용자에게 로또 번호 입력받은 후 입력값 검증
