+++ src/constant/lottoConstants.js
@@ -4,11 +4,16 @@ export const MESSAGE = {
 
 export const PROPT_MESSAGE = {
   INPUT_PURCHASE_AMOUNT: '구입금액을 입력해 주세요.',
+  INPUT_WINNING_NUMBER: '로또 당첨 번호를 입력해주세요.',
 };
 
 export const ERROR_MESSAGE = {
   NOT_A_NUMBER: '[ERROR] 구입금액은 숫자로 입력해야 합니다.',
   NOT_DIVISIBLE: '[ERROR] 로또는 1,000원 단위로 구매 가능합니다.',
+  LESS_LOTTO_NUMBERS: '[ERROR] 로또 번호는 6개여야 합니다.',
+  LOTTO_NUMBER_HAVE_DUPLICATE: '[ERROR] 중복된 당첨 번호가 있습니다.',
+  MUST_BE_NUMBER: '[ERROR] 로또 번호는 숫자여야 합니다.',
+  INVALID_LOTTO_NUMBER_RANGE: '[ERROR] 로또 번호는 1부터 45 사이의 숫자여야 합니다.',
 };
 
 export const PRICE = {
